pyproject.toml
[tool.poetry]
name = "tqdne"
version = "0.1.0"
readme = "README.rst"
description = ""
authors = ["Your Name <you@example.com>"]


[tool.poetry.dependencies]
python = ">=3.10.0,<4.0"
tqdm = "*"
scikit-learn = "*"
scikit-image = "*"
numpy = "*"
matplotlib = "*"
python-dotenv = "*"
setuptools = "*"
wandb = "*"
torch = "*"
torchvision = "*"
pytorch-lightning = ">=2.0.0"
h5py = "*"


[tool.poetry.dev-dependencies]
pytest = "*"
jupyterlab = "*"
twine = "*"
setuptools = "*"
wheel = "*"
flake8 = "*"
pylint = "*"
sphinx = "<7.0.0"
numpydoc = "*"
sphinxcontrib-bibtex = "*"
sphinx-gallery = "*"
memory_profiler = "*"
sphinx-rtd-theme = "*"
sphinx-copybutton = "*"
nbsphinx = "*"
pandoc = "*"
black = "*"

[tool.poetry.group.dev.dependencies]
notebook = "^6.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


